Refactor tool_util only used by tool_shed code into tool_shed.
