+++ lib/galaxy/tool_shed/util/tool_util.py
@@ -1,17 +1,12 @@
-import logging
 import os
 import shutil
 from typing import Optional
 
-import galaxy.tools
 from galaxy import util
 from galaxy.datatypes.sniff import is_column_based
 from galaxy.tool_shed.util import basic_util
 from galaxy.util import checkers
-from galaxy.util.expressions import ExpressionContext
 from galaxy.web.form_builder import SelectField
-
-log = logging.getLogger(__name__)
 
 
 def build_shed_tool_conf_select_field(app):
@@ -197,34 +192,6 @@ def _is_data_index_sample_file(file_path):
     return True
 
 
-def new_state(trans, tool, invalid=False):
-    """Create a new `DefaultToolState` for the received tool.  Only inputs on the first page will be initialized."""
-    state = galaxy.tools.DefaultToolState()
-    state.inputs = {}
-    if invalid:
-        # We're attempting to display a tool in the tool shed that has been determined to have errors, so is invalid.
-        return state
-    try:
-        # Attempt to generate the tool state using the standard Galaxy-side code
-        return tool.new_state(trans)
-    except Exception as e:
-        # Fall back to building tool state as below
-        log.debug(
-            'Failed to build tool state for tool "%s" using standard method, will try to fall back on custom method: %s',
-            tool.id,
-            e,
-        )
-    inputs = tool.inputs_by_page[0]
-    context = ExpressionContext(state.inputs, parent=None)
-    for input in inputs.values():
-        try:
-            state.inputs[input.name] = input.get_initial_value(trans, context)
-        except Exception:
-            # FIXME: not all values should be an empty list
-            state.inputs[input.name] = []
-    return state
-
-
 def panel_entry_per_tool(tool_section_dict):
     # Return True if tool_section_dict looks like this.
     # {<Tool guid> :
@@ -251,6 +218,5 @@ __all__ = (
     "copy_sample_files",
     "generate_message_for_invalid_tools",
     "handle_missing_index_file",
-    "new_state",
     "panel_entry_per_tool",
 )
